% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionFunctions.R
\name{avg_conf_pred}
\alias{avg_conf_pred}
\title{Obtain a distance-from-average conformal prediction interval}
\usage{
avg_conf_pred(Y, alpha)
}
\arguments{
\item{Y}{Observed data vector}

\item{alpha}{Prediction error rate}
}
\value{
alpha-level prediction interval
}
\description{
This function computes a conformal prediction region under the distance-from-average
non-conformity measure. That is, |a + bz*| <= |ci + di z^*| where i indexes training data.
}
