% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtaPrediction.R
\name{dtaPrediction}
\alias{dtaPrediction}
\title{Obtain a distance-to-average conformal prediction interval}
\usage{
dtaPrediction(Y, alpha = 0.15)
}
\arguments{
\item{Y}{Observed data vector}

\item{alpha}{Prediction error rate}
}
\value{
pred object
}
\description{
This function computes a conformal prediction region under the distance-from-average
non-conformity measure. That is, |a + bz*| <= |ci + di z^*| where i indexes training data.
}
