% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBayesFunctionsContinuous.R
\name{EB_values}
\alias{EB_values}
\title{Obtain empirical Bayesian estimates for group j}
\usage{
EB_values(j, Y, group, W = NA, X = NA)
}
\arguments{
\item{j}{Obtain EB values for group in index j- numeric value in group}

\item{Y}{Data vector}

\item{group}{index vecter of the same lenght as Y}

\item{W}{Non-standardized adjacency matrix}

\item{X}{Group-level covariates}
}
\value{
empirical Bayesian estimates of population mean and it's variance
}
\description{
This function returns empirical Bayesian estimates for a specified group from
the conjugate normal spatial Fay-Herriot model.
}
