% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionFunctions.R
\name{fab_conf_pred}
\alias{fab_conf_pred}
\title{Obtain a FAB conformal prediction interval}
\usage{
fab_conf_pred(Y, mu = 0, tau2 = 1, alpha = 0.05)
}
\arguments{
\item{Y}{Observed data vector}

\item{mu}{Prior expected mean of the population mean}

\item{tau2}{Prior expected variance of the population mean}

\item{alpha}{Prediction error rate}
}
\value{
alpha-level prediction interval
}
\description{
This function computes a FAB conformal prediction region as described in
Bersson and Hoff 2022.
}
