% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabPrediction-package.R
\docType{package}
\name{fabPrediction}
\alias{fabPrediction}
\title{fabPrediction: Compute FAB Conformal Prediction Intervals}
\description{
A package for computing and plotting
prediction intervals for numerical data or prediction
sets for categorical data using prior information. 
Empirical Bayes procedures to estimate the prior information
from multi-group data are included.
}
\references{
E. Bersson and P.D. Hoff. (2023) Frequentist Prediction Sets for 
Species Abundance using Indirect Information. Preprint.

E. Bersson and P.D. Hoff. (2023) Optimal Conformal Prediction for Small Areas.
Journal of Survey Statistics and Methodology, forthcoming.
}
\author{
Elizabeth Bersson, \cr Maintainer: Elizabeth Bersson <elb75@duke.edu>
}
\concept{pred methods}
\keyword{package}
\keyword{prediction}
