% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBayesFunctionsCategorical.R
\name{init.MoM}
\alias{init.MoM}
\title{Obtain inital guess of MLE of the marginal Dirichlet-multinomial likelihood}
\usage{
init.MoM(D)
}
\arguments{
\item{D}{matrix (JxK) of counts; each row is a sample from a MN distribution with K categories}
}
\value{
Hessian
}
\description{
Method of moment matching to obtain an initial guess of the MLE, as in Minka (2000).
}
