% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalPrediction.R
\name{normalPrediction}
\alias{normalPrediction}
\title{Obtain a pivot prediction interval}
\usage{
normalPrediction(Y, alpha = 0.15)
}
\arguments{
\item{Y}{Observed data vector}

\item{alpha}{Prediction error rate}
}
\value{
pred object
}
\description{
This function computes a prediction interval under assumed normality.
}
