% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBayesFunctionsContinuous.R
\name{plugin_values}
\alias{plugin_values}
\title{Obtain empirical Bayesian estimates for conjugate normal spatial Fay-Herriot model}
\usage{
plugin_values(Y, group, W = NA, X = NA)
}
\arguments{
\item{Y}{Data vector}

\item{group}{Group membership of each entry in Y}

\item{W}{Adjacency matrix}

\item{X}{Group-level covariates}
}
\value{
plug-in values of spatial Fay-Herriot model
}
\description{
This function returns plug-in values for a conjugate
normal spatial Fay-Herriot model.
}
