% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyaGradient.R
\name{polyaGradient}
\alias{polyaGradient}
\title{Obtain gradient of the marginal Dirichlet-multinomial likelihood}
\usage{
polyaGradient(D, gamma, Nj = rowSums(D), K = ncol(D))
}
\arguments{
\item{D}{matrix (JxK) of counts; each row is a sample from a MN distribution with K categories}

\item{gamma}{current value of prior concentration parameter}

\item{Nj}{sample sizes of the J groups}

\item{K}{number of categories}
}
\value{
gradient
}
\description{
Obtain gradient of the marginal Dirichlet-multinomial likelihood
}
