% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyaHessian.R
\name{polyaHessian}
\alias{polyaHessian}
\title{Obtain Hessian of the marginal Dirichlet-multinomial likelihood}
\usage{
polyaHessian(D, gamma, Nj = rowSums(D), K = ncol(D))
}
\arguments{
\item{D}{matrix (JxK) of counts; each row is a sample from a MN distribution with K categories}

\item{gamma}{current value of prior concentration parameter}

\item{Nj}{sample sizes of the J groups}

\item{K}{number of categories}
}
\value{
Hessian
}
\description{
Obtain Hessian of the marginal Dirichlet-multinomial likelihood
}
