% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radon}
\alias{radon}
\title{Minnesota Radon Data}
\format{
A matrix.
}
\source{
\href{http://www.stat.columbia.edu/~gelman/arm/examples/radon/}{ARM Data}
}
\usage{
data(radon)
}
\description{
Data from a national US EPA survey of household radon values. County index contained in group column.
}
\references{
US Environmental Protection Agency (1992) National residential radon survey: summary report.
Washington, DC; DOI EPA402-R-92-011.
}
\keyword{datasets}
