% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate a random sample from a Dirichlet distribution}
\usage{
rdirichlet(gamma)
}
\arguments{
\item{gamma}{Prior concentration vector of length K}
}
\value{
a vector of length K that is a random sample from a Dirichlet distribution
}
\description{
Generate a random sample from a Dirichlet distribution
}
