% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_standardize.R
\name{row_standardize}
\alias{row_standardize}
\title{Row standardize a matrix}
\usage{
row_standardize(W)
}
\arguments{
\item{W}{matrix}
}
\value{
row-standardized matrix
}
\description{
Row standardize a matrix
}
