% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W}
\alias{W}
\title{Minnesota County Adjacency Matrix}
\format{
A matrix.
}
\usage{
data(W)
}
\description{
Adjacency matrix for MN counties based on group index that matches radon data.
}
\keyword{datasets}
