% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pred}
\alias{plot.pred}
\title{Plot a \code{pred} object}
\usage{
\method{plot}{pred}(obj, main = "", xlab = "", cex.axis = 1)
}
\arguments{
\item{obj}{pred object- a list classified as pred containing objects data and bound}

\item{plot.all}{(TO BE ADDED) T displays all categories when plotting a categorical prediction set.}
}
\value{
capability to plot pred object. More details: the command \code{plot(obj)} returns a line plot with tick marks denoting observed data points and a red band denoting the prediction interval
}
\description{
Plot a \code{pred} object
}
