% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plot.pred}
\alias{plot.pred}
\title{Plot a \code{pred} object}
\usage{
\method{plot}{pred}(obj, main = "", xlab = "")
}
\arguments{
\item{obj}{pred object- a list classified as pred containing objects data and bound}

\item{main}{plot title}

\item{xlab}{xaxis title}
}
\value{
capability to plot pred object. More details: the command \code{plot(obj)} returns a line plot with tick marks denoting observed data points and a red band denoting the prediction interval
}
\description{
Plot a \code{pred} object
}
